refactor: Return public API from activate function
This commit refactors the `activate` function in `extension.ts` to return a `PublicApi` object. This object provides two methods: `getMode` and `setMode`. The `getMode` method retrieves the current mode, while the `setMode` method allows setting the mode. This change improves the extensibility of the extension and provides a cleaner interface for interacting with the extension's functionality.